requirements.txt
easygui==0.98.3
pycryptodomex==3.18.0
tqdm==4.64.1
